chore: renamed folders, files, relocated files
