+++ include/color.h
@@ -1,0 +1,12 @@
+#ifndef COLOR_H
+#define COLOR_H
+
+#define RESET "\033[0m"
+#define BOLD "\033[1m"
+#define CYAN "\033[36m"
+#define GREEN "\033[0;32m"
+#define RED "\033[0;31m"
+#define GREY "\033[90m"
+#define BLUE "\033[0;34m"
+
+#endif
