+++ include/http-response.h
@@ -4,8 +4,7 @@
 # include <iostream>
 # include <map>
 # include <ctime>
-
-// #include "sockets-polling.h"
+#include "color.h"
 #include "Configuration.hpp"
 
 class Server;
