+++ include/read-request.h
@@ -2,22 +2,17 @@
 # define READ_REQUEST_H
 
 #include <sys/socket.h>
+#include <utility>
+#include <map>
 
 #include "Configuration.hpp"
 #include "Client.h"
+#include "color.h"
 
 #define HEADER_END "\r\n\r\n"
 #define CRLF "\r\n"
 #define CRLF_LENGTH 2
 #define DOUBLE_CRLF_LENGTH 4
-
-#define RESET "\033[0m"
-#define BOLD "\033[1m"
-#define CYAN "\033[36m"
-#define GREEN "\033[0;32m"
-#define RED "\033[0;31m"
-#define GREY "\033[90m"
-#define BLUE "\033[0;34m"
 
 enum readReturnVal {
     RECV_OK = 0,
@@ -32,5 +27,6 @@ int readRequestBody(Client& client, std::string& bodyStr, const size_t bufferSiz
 int readByContentLength(Client& client, std::string& bodyStr, const size_t bufferSize, const size_t maxSize);
 int readByChunkedEncoding(Client& client, std::string& bodyStr, const size_t bufferSize, const size_t maxSize);
 int readFromSocket(Client& client, int bufferSize);
+void dispatchRequest(Client& client);
 
 #endif
