+++ include/utils.h
@@ -2,8 +2,11 @@
 # define UTILS_H
 
 # include <iostream>
+# include "Configuration.hpp"
 
 std::string toLower(const std::string& str);
 bool isDigitsOnly(const std::string& str);
+
+
 
 #endif
