+++ src/handle-sockets/disconnect-client.cpp
@@ -1,4 +1,4 @@
-#include "sockets-polling.h"
+#include "handle-sockets.h"
 
 std::vector<Client>::iterator disconnectClient(std::vector<Client>& clients, std::vector<Client>::iterator &clientIt, std::vector<struct pollfd>& pfds)
 {
