+++ src/handle-sockets/setup-listeners.cpp
@@ -1,5 +1,5 @@
-#include "../../include/sockets-polling.h"
 #include <string.h> // memset
+#include "handle-sockets.h"
 
 // Return a listening socket
 static int getListenerSocket(const std::string& host, const std::string& port)
