+++ src/http-response/get-handler.cpp
@@ -1,8 +1,10 @@
-#include "http-response.h"
-#include "http-exception.h"
-#include "sockets-polling.h"
 #include <sys/stat.h>
 #include <dirent.h> //for opendir, readdir, closedir
+#include <fcntl.h>
+#include <unistd.h>
+#include "http-response.h"
+#include "http-exception.h"
+#include "Client.h"
 
 static std::string readDirectorytoString(const std::string &directoryPath, const std::string& uri);
