+++ src/main.cpp
@@ -1,6 +1,7 @@
-#include "sockets-polling.h"
 #include "Configuration.hpp"
 #include "timeout.h"
+#include "handle-sockets.h"
+#include "poll-loop.h"
 #include <map>
 
 std::string getConfigFileString(int argc, char **argv)
