+++ src/poll-loop/poll-loop-utils.cpp
@@ -1,4 +1,5 @@
-#include "sockets-polling.h"
+#include <algorithm> // for isListener
+#include "poll-loop.h"
 
 bool isListener(std::vector<int>& listeners, int fd)
 {
