+++ src/timeout/handle-timeout.cpp
@@ -1,5 +1,6 @@
 #include "timeout.h"
-#include "sockets-polling.h"
+#include "read-request.h"
+#include "handle-sockets.h"
 
 time_t getNowInSeconds() {
     struct timeval tv;
